requirements.txt
# Multi-Agent Employee Relocation Concierge
# uAgents + Chat Protocol + Payment Protocol (ASI:One / Agentverse / Stripe)
uagents>=0.24.0
uagents-core>=0.4.3
openai>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
stripe>=10.0.0

# Brotli decompression for Agentverse (content-encoding: br)
# Fixes: ClientPayloadError "Can not decode content-encoding: br" and
#        TypeError: process() takes exactly 1 argument (2 given)
brotlicffi>=1.2.0
